Added visual test, describe and before/after each hooks
